ウィンドウ上のテキストに味方や敵、技の配列が反映されるように修正

+++ Assets/Scripts/Item/Weapon.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class Weapon : MonoBehaviour
 {
-    private int attackPower;
+    private float attackPower;
 
 }
